refactor: 각 컴포넌트 폴더 및 파일의 이름을 카멜에서 케밥 케이스로 변경
