+++ src/components/pages/group/form/index.ts
@@ -1,2 +1,2 @@
-export * from './CreateGroupForm';
-export * from './EditGroupForm';
+export * from './create-group-form';
+export * from './edit-group-form';
