--- src/components/pages/profile/BasicInformationForm/components/index.ts
@@ -1,2 +1,0 @@
-export * from './RadioGroup';
-export * from './ImageUpload';
