+++ src/components/pages/profile/basic-information-form/components/index.ts
@@ -1,0 +1,2 @@
+export * from './radio-group';
+export * from './image-upload';
