+++ src/components/pages/profile/index.ts
@@ -1,5 +1,5 @@
-export * from './Header';
-export * from './BasicInformationForm';
-export * from './AddtionalInformationForm';
-export * from './CreateProfileComplete';
-export * from './Stepper';
+export * from './header';
+export * from './basic-information-form';
+export * from './additional-information-form';
+export * from './create-profile-complete';
+export * from './stepper';
